simdnbt

Module borrow

The borrowed variant of NBT. This is useful if you’re only reading data and you can keep a reference to the original buffer.

Structs

BaseNbt

A complete NBT container. This contains a name and a compound tag.

BaseNbtCompound

A nameless NBT container. This only contains a compound tag. This contains a TagAllocator, so it can exist independently from a BaseNbt.

BaseNbtTag

A nameless NBT tag.

NbtCompound

NbtCompoundIter

NbtCompoundList

A wrapper over NbtCompoundListIter that acts more like a Vec.

NbtCompoundListIter

NbtList

A list of NBT tags of a single type.

NbtListList

A wrapper over NbtListListIter that acts more like a Vec.

NbtListListIter

An iterator over a list of lists.

NbtTag

Enums

Nbt

Either a complete NBT container, or nothing.

Functions

read

Read a normal root NBT compound. This is either empty or has a name and compound tag.

read_compound

Read a compound tag. This may have any number of items.

read_optional_tag

Read any NBT tag, without reading its name. This may be any type of tag, including an end tag.

read_tag

Read an NBT tag, without reading its name. This may be any type of tag except for an end tag. If you need to be able to handle end tags, use read_optional_tag.

read_unnamed

Read a root NBT compound, but without reading the name. This is used in Minecraft when reading NBT over the network.