simdnbt

Module owned

The owned variant of NBT. This is useful if you’re writing NBT or if you can’t keep a reference to the original data.

Structs

BaseNbt

A complete NBT container. This contains a name and a compound tag.

NbtCompound

A list of named tags. The order of the tags is preserved.

Enums

Nbt

NbtList

A list of NBT tags of a single type.

NbtTag

A single NBT tag.

Functions

read

Read a normal root NBT compound. This is either empty or has a name and compound tag.

read_compound

Read a compound tag. This may have any number of items.

read_optional_tag

Read any NBT tag, without reading its name. This may be any type of tag, including an end tag.

read_tag

Read an NBT tag, without reading its name. This may be any type of tag except for an end tag. If you need to be able to handle end tags, use read_optional_tag.

read_unnamed

Read a root NBT compound, but without reading the name. This is used in Minecraft when reading NBT over the network.