simdnbt::owned

Struct NbtCompound

pub struct NbtCompound { /* private fields */ }

A list of named tags. The order of the tags is preserved.

Implementations

impl NbtCompound

pub fn new() -> Self

pub fn from_values(values: Vec<(Mutf8String, NbtTag)>) -> Self

pub fn write(&self, data: &mut Vec<u8>)

pub fn get(&self, name: &str) -> Option<&NbtTag>

pub fn get_mut(&mut self, name: &str) -> Option<&mut NbtTag>

pub fn take(&mut self, name: &str) -> Option<NbtTag>

Get an owned tag from the compound by swapping it with a dummy tag.

pub fn contains(&self, name: &str) -> bool

Returns whether there is a tag with the given name.

pub fn byte(&self, name: &str) -> Option<i8>

pub fn byte_mut(&mut self, name: &str) -> Option<&mut i8>

pub fn short(&self, name: &str) -> Option<i16>

pub fn short_mut(&mut self, name: &str) -> Option<&mut i16>

pub fn int(&self, name: &str) -> Option<i32>

pub fn int_mut(&mut self, name: &str) -> Option<&mut i32>

pub fn long(&self, name: &str) -> Option<i64>

pub fn long_mut(&mut self, name: &str) -> Option<&mut i64>

pub fn float(&self, name: &str) -> Option<f32>

pub fn float_mut(&mut self, name: &str) -> Option<&mut f32>

pub fn double(&self, name: &str) -> Option<f64>

pub fn double_mut(&mut self, name: &str) -> Option<&mut f64>

pub fn byte_array(&self, name: &str) -> Option<&[u8]>

pub fn byte_array_mut(&mut self, name: &str) -> Option<&mut Vec<u8>>

pub fn string(&self, name: &str) -> Option<&Mutf8Str>

pub fn string_mut(&mut self, name: &str) -> Option<&mut Mutf8String>

pub fn list(&self, name: &str) -> Option<&NbtList>

pub fn list_mut(&mut self, name: &str) -> Option<&mut NbtList>

pub fn compound(&self, name: &str) -> Option<&NbtCompound>

pub fn compound_mut(&mut self, name: &str) -> Option<&mut NbtCompound>

pub fn int_array(&self, name: &str) -> Option<&[i32]>

pub fn int_array_mut(&mut self, name: &str) -> Option<&mut Vec<i32>>

pub fn long_array(&self, name: &str) -> Option<&[i64]>

pub fn long_array_mut(&mut self, name: &str) -> Option<&mut Vec<i64>>

pub fn iter(&self) -> impl Iterator<Item = (&Mutf8Str, &NbtTag)>

pub fn iter_mut(&mut self) -> impl Iterator<Item = (&Mutf8Str, &mut NbtTag)>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn keys(&self) -> impl Iterator<Item = &Mutf8Str>

pub fn keys_mut(&mut self) -> impl Iterator<Item = &mut Mutf8String>

pub fn values(&self) -> impl Iterator<Item = &NbtTag>

pub fn values_mut(&mut self) -> impl Iterator<Item = &mut NbtTag>

pub fn clear(&mut self)

pub fn insert(&mut self, name: impl Into<Mutf8String>, tag: impl ToNbtTag)

pub fn extend( &mut self, other: impl IntoIterator<Item = (impl Into<Mutf8String>, impl ToNbtTag)>, )

pub fn remove(&mut self, name: &str) -> Option<NbtTag>